% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{left_join2}
\alias{left_join2}
\alias{inner_join2}
\alias{right_join2}
\alias{full_join2}
\title{Perform dplyr joins with incompatible types}
\usage{
left_join2(
  x,
  y,
  by = NULL,
  coerce_on_conflict = c("character", "numeric"),
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

inner_join2(
  x,
  y,
  by = NULL,
  coerce_on_conflict = c("character", "numeric"),
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

right_join2(
  x,
  y,
  by = NULL,
  coerce_on_conflict = c("character", "numeric"),
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)

full_join2(
  x,
  y,
  by = NULL,
  coerce_on_conflict = c("character", "numeric"),
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE
)
}
\arguments{
\item{x}{A data frame}

\item{y}{A data frame}

\item{by}{A character vector of variables to join by. Can be NULL.}

\item{coerce_on_conflict}{Either 'character' or 'numeric'.}

\item{suffix}{If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.}

\item{...}{Other parameters passed on to methods}

\item{keep}{Should the join keys from both x and y be preserved in the output?}
}
\value{
a data frame
}
\description{
These joins will coerce key columns to a common atomic type.
}
\examples{
df1 <- data.frame(x = 1:10, b = 1:10, y = letters[1:10])
df2 <- data.frame(x = as.character(1:10), z = letters[11:20])
left_join2(df1, df2)
}
