% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{filter_pattern}
\alias{filter_pattern}
\title{Grep and filter a data frame by pattern}
\usage{
filter_pattern(.data, col, pattern, ...)
}
\arguments{
\item{.data}{a table of data.}

\item{col}{a variable.}

\item{pattern}{string containing a regular expression to be matched in the given character vector.}

\item{...}{additional arguments passed to grepl}
}
\description{
Grep and filter a data frame by pattern
}
