% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_na.R
\name{keep_na}
\alias{keep_na}
\title{Keep rows containing missing values}
\usage{
keep_na(.data, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{A selection of columns.}
}
\description{
Keep rows containing missing values
}
