% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{keep_pattern}
\alias{keep_pattern}
\alias{discard_pattern}
\title{Grep, keep or discard a list or vector by pattern}
\usage{
keep_pattern(x, pattern, ...)

discard_pattern(x, pattern, ...)
}
\arguments{
\item{x}{a list of vector}

\item{pattern}{string containing a regular expression to be matched in the given character vector.}

\item{...}{additional arguments passed to grepl}
}
\description{
Grep, keep or discard a list or vector by pattern
}
