% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_row_values.R
\name{shift_row_values}
\alias{shift_row_values}
\title{Shift row values left or right}
\usage{
shift_row_values(.data, .dir = "left", at = NULL)
}
\arguments{
\item{.data}{a table of data.}

\item{.dir}{the shift direction as a string, one of "left" or "right"}

\item{at}{the row indices at which to shift}
}
\value{
a tibble
}
\description{
Shift row values left or right
}
