% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R
\name{cast_character}
\alias{cast_character}
\alias{cast_numeric}
\alias{cast_logical}
\title{Casting functions}
\usage{
cast_character(data, ...)

cast_numeric(data, ...)

cast_logical(data, ...)
}
\arguments{
\item{data}{a table of data}

\item{...}{columns}
}
\description{
Casting functions
}
