% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_na.R
\name{keep_na}
\alias{keep_na}
\title{Keep rows containing missing values}
\usage{
keep_na(.data, ..., .logic = c("AND", "OR"))
}
\arguments{
\item{.data}{A table of data.}

\item{...}{A selection of columns. If empty, all columns are selected.}

\item{.logic}{boolean, either 'AND' or 'OR'. Logic for keeping NAs.}
}
\value{
A data frame.
}
\description{
Keep rows containing missing values
}
\examples{
library(dplyr)
df <- tibble(x = c(1, 2, NA, NA), y = c("a", NA, "b", NA))
df \%>\% keep_na()
df \%>\% keep_na(x)

vars <- "y"
df \%>\% keep_na(x, any_of(vars))
}
