% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{filter_split}
\alias{filter_split}
\alias{select_split}
\alias{mutate_split}
\alias{distinct_split}
\alias{transmute_split}
\alias{slice_split}
\title{Perform various operations before splitting}
\usage{
filter_split(.data, ...)

select_split(.data, ...)

mutate_split(.data, ...)

distinct_split(.data, ..., simplify = TRUE)

transmute_split(.data, ..., simplify = TRUE)

slice_split(.data, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Expressions to be evaluated.}

\item{simplify}{Boolean, whether to unlist the returned split.}
}
\description{
Learn more in \code{vignette("hacksaw-splitting")}.
}
