% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{var_max}
\alias{var_max}
\title{Return the indices of n max values of a variable}
\usage{
var_max(var, n = 6)
}
\arguments{
\item{var}{the variable to use.}

\item{n}{number of rows to return.}
}
\description{
Return the indices of n max values of a variable
}
