% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{var_min}
\alias{var_min}
\title{Return the indices of n min values of a variable}
\usage{
var_min(var, n = 6)
}
\arguments{
\item{var}{the variable to use.}

\item{n}{number of rows to return.}
}
\description{
Return the indices of n min values of a variable
}
\examples{
var_min(1:10)
}
